Rename `LeftSidebar` to `NavigationBar` and extract.
