+++ client/src/main/kotlin/io/spine/examples/chatspn/desktop/navigation/Navigation.kt
@@ -27,45 +27,21 @@
 package io.spine.examples.chatspn.desktop.navigation
 
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.Surface
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Menu
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.drawBehind
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
 import io.spine.examples.chatspn.desktop.DesktopClient
 import io.spine.examples.chatspn.desktop.LoginPage
 import io.spine.examples.chatspn.desktop.RegistrationPage
 import io.spine.examples.chatspn.desktop.chat.ChatPage
 import io.spine.examples.chatspn.desktop.profile.ProfilePage
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
 
 /**
  * Displays the current page of the application.
@@ -94,11 +70,11 @@ public fun CurrentPage(client: DesktopClient) {
             }
         )
         Page.CHAT -> Row {
-            LeftSidebar(model)
+            NavigationBar(model)
             ConfiguredChatPage(client, model)
         }
         Page.PROFILE -> Row {
-            LeftSidebar(model)
+            NavigationBar(model)
             ConfiguredProfilePage(client, model)
         }
     }
@@ -158,122 +134,4 @@ private fun ConfiguredProfilePage(client: DesktopClient, model: NavigationModel)
         { model.currentPage.value = Page.REGISTRATION },
         { model.selectPersonalChat(it) }
     )
-}
-
-/**
- * Displays the left sidebar.
- */
-@Composable
-private fun LeftSidebar(model: NavigationModel) {
-    Surface(
-        Modifier
-            .padding(end = 1.dp)
-            .drawBehind {
-                drawLine(
-                    color = Color.Gray,
-                    start = Offset(size.width, 0f),
-                    end = Offset(size.width, size.height),
-                    strokeWidth = 1.dp.toPx()
-                )
-            },
-    ) {
-        Column(
-            Modifier
-                .fillMaxHeight()
-                .width(256.dp)
-        ) {
-            Row(
-                Modifier
-                    .fillMaxWidth()
-                    .height(54.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Spacer(Modifier.width(10.dp))
-                MenuButton(model)
-                Spacer(Modifier.width(14.dp))
-                UserSearchField(model)
-            }
-            ChatList(model)
-        }
-    }
-}
-
-/**
- * Displays the menu button.
- */
-@Composable
-private fun MenuButton(model: NavigationModel) {
-    val currentPage = remember { model.currentPage }
-    val isUserProfileOpen = currentPage.value == Page.PROFILE
-    val isAuthenticatedUser = model.authenticatedUser
-        .equals(model.profilePageState.userProfile.value)
-    val containerColor = if (isUserProfileOpen && isAuthenticatedUser)
-        MaterialTheme.colorScheme.inverseSurface
-    else
-        MaterialTheme.colorScheme.surface
-    Button(
-        modifier = Modifier.size(42.dp),
-        onClick = {
-            if (isUserProfileOpen && isAuthenticatedUser) {
-                model.currentPage.value = Page.CHAT
-                model.profilePageState.clear()
-            } else {
-                model.openUserProfile(model.authenticatedUser.id)
-            }
-        },
-        shape = CircleShape,
-        colors = ButtonDefaults.buttonColors(
-            containerColor = containerColor
-        ),
-        contentPadding = PaddingValues(6.dp)
-    ) {
-        Icon(
-            Icons.Default.Menu,
-            "Menu",
-            Modifier.size(20.dp),
-            MaterialTheme.colorScheme.onBackground
-        )
-    }
-}
-
-/**
- * Displays the input field to find the user.
- */
-@Composable
-private fun UserSearchField(model: NavigationModel) {
-    val viewScope = rememberCoroutineScope { Dispatchers.Default }
-    val inputText = remember { model.userSearchFieldState.userEmailState }
-    val onSearch = {
-        val email = inputText.value
-        viewScope.launch {
-            if (email.trim().isNotEmpty()) {
-                model.createPersonalChat(email.trim())
-            }
-        }
-        inputText.value = ""
-    }
-    SearchField(inputText, onSearch)
-}
-
-/**
- * Displays the list of chat previews.
- */
-@Composable
-private fun ChatList(model: NavigationModel) {
-    val chats by model.chats().collectAsState()
-    val selectedChat = model.selectedChat
-    LazyColumn(
-        Modifier.fillMaxSize()
-    ) {
-        chats.forEachIndexed { index, chat ->
-            item(key = index) {
-                ChatPreviewPanel(
-                    chat,
-                    chat.id.equals(selectedChat.value)
-                ) {
-                    model.selectChat(chat.id)
-                }
-            }
-        }
-    }
 }
